+++ MachineLearning-master/SVM/SVM_by_SMO/testSVM-SMO.py
@@ -16,18 +16,18 @@ SVM SMO 测试脚本
 import csv, os, sys
 import numpy as np
 from SVCSMO import SVCSMO
-filepath = os.path.dirname(os.path.abspath(__file__))
+filepath = os.path.dirname(os.path.abspath(__file__))  # # 当前脚本所在目录，用于拼接数据文件路径
 
 def readData(filename, header=True):
     """读取 CSV 文件，返回 (数据矩阵, 表头)。"""
     data, header = [], None
     with open(filename, 'r') as csvfile:
-        spamreader = csv.reader(csvfile, delimiter=',')
+        spamreader = csv.reader(csvfile, delimiter=',')  # # 逗号分隔读取
         if header:
-            header = next(spamreader)
+            header = next(spamreader)  # # 有表头时先读一行表头
         for row in spamreader:
             data.append(row)
-    return (np.array(data), np.array(header))
+    return (np.array(data), np.array(header))  # # 数据转为二维数组
 
 def calc_acc(y, y_hat):
     """计算准确率：分别统计预测为 1 和 -1 时的正确数。"""
@@ -44,14 +44,14 @@ def calc_mse(y, y_hat):
 def test_main(filename='data/iris-virginica.txt', C=1.0, kernel_type='linear', epsilon=0.001):
     """SVM SMO 完整测试流程：加载数据 -> 训练 -> 预测 -> 输出指标。"""
     # Load data
-    (data, _) = readData('%s/%s' % (filepath, filename), header=False)
-    data = data.astype(float)
+    (data, _) = readData('%s/%s' % (filepath, filename), header=False)  # # 读取 CSV（无表头），路径拼脚本目录
+    data = data.astype(float)  # # 字符串转浮点
 
     # 划分特征和标签
     X, y = data[:,0:-1], data[:,-1].astype(int)
 
     # 初始化 SMO SVM 模型
-    model = SVCSMO()
+    model = SVCSMO()  # # 使用默认参数：线性核、C=1.0
 
     # 训练模型，返回支持向量数组和迭代次数
     support_vectors, iterations = model.fit(X, y)
@@ -63,9 +63,10 @@ def test_main(filename='data/iris-virginica.txt', C=1.0, kernel_type='linear', e
     y_hat = model.predict(X)
 
     # Calculate accuracy
-    acc = calc_acc(y, y_hat)
-    mse = calc_mse(y, y_hat)
+    acc = calc_acc(y, y_hat)  # # 准确率 = 正确预测数 / 样本总数
+    mse = calc_mse(y, y_hat)  # # 均方误差
 
+    # # 输出模型参数与评估指标
     print("Support vector count: %d" % (sv_count))
     print("bias:\t\t%.3f" % (model.b))
     print("w:\t\t" + str(model.w))
@@ -74,12 +75,12 @@ def test_main(filename='data/iris-virginica.txt', C=1.0, kernel_type='linear', e
     print("Converged after %d iterations" % (iterations))
 
 if __name__ == '__main__':
-    param = {}
+    param = {}  # # 测试参数
     param['filename'] = './small_data/iris-slwc.txt'
-    param['C'] = 0.1
+    param['C'] = 0.1  # # 较小的 C：软间隔容忍部分误分
     param['kernel_type'] = 'linear'
     param['epsilon'] = 0.001
 
 
-    test_main(**param)
+    test_main(**param)  # # 运行完整测试流程
